+++ inventory/src/main/java/com/emall/inventory/repository/InventoryItemEntity.java
@@ -17,13 +17,13 @@ public class InventoryItemEntity {
     private Long skuId;
 
     @TableField("total")
-    private Integer total;
+    private Long total;
 
     @TableField("reserved")
-    private Integer reserved;
+    private Long reserved;
 
     @TableField("sold")
-    private Integer sold;
+    private Long sold;
 
     @TableField(value = "updated_at", fill = FieldFill.INSERT_UPDATE)
     private LocalDateTime updatedAt;
